+++ cypress/integration/Moderation.ReportContent.feature
@@ -8,51 +8,47 @@ Feature: Report and Moderate
   So I can look into it and decide what to do
 
   Background:
-    Given we have the following user accounts:
-      | id            | name                                          |
-      | u67           | David Irving                                  |
-      | annoying-user | I'm gonna mute Moderators and Admins HA HA HA |
-
-    Given we have the following posts in our database:
+    Given the following "users" are in the database:
+      | slug      | email                 | password | id            | name                                          | role      | termsAndConditionsAgreedVersion |
+      | user      | user@example.org      | abcd     | user          | User-Chad                                     | user      | 0.0.4                           |
+      | moderator | moderator@example.org | 1234     | moderator     | Mod-Man                                       | moderator | 0.0.4                           |
+      | annoying  | annoying@example.org  | 1234     | annoying-user | I'm gonna mute Moderators and Admins HA HA HA | user      | 0.0.4                           |
+    And the following "posts" are in the database:
       | authorId      | id | title                         | content                                              |
-      | u67           | p1 | The Truth about the Holocaust | It never existed!                                    |
+      | annoying-user | p1 | The Truth about the Holocaust | It never existed!                                    |
       | annoying-user | p2 | Fake news                     | This content is demonstratably infactual in some way |
+
   Scenario Outline: Report a post from various pages
-    Given I am logged in with a "user" role
-    When I see David Irving's post on the <Page>
+    When I am logged in as "user"
+    And I navigate to page "<Page>" 
     And I click on "Report Post" from the content menu of the post
     And I confirm the reporting dialog because it is a criminal act under German law:
       """
       Do you really want to report the contribution "The Truth about the Holocaust"?
       """
-    Then I see a success message:
-      """
-      Thanks for reporting!
-      """
+    Then I see a toaster with "Thanks for reporting!"
     Examples:
-      | Page         |
-      | newsfeed page|
-      | post page    |
+      | Page     |
+      | /        |
+      | /post/p1 |
 
   Scenario: Report user
-    Given I am logged in with a "user" role
-    And I see David Irving's post on the post page
+    Given I am logged in as "user"
+    And I navigate to page "/post/the-truth-about-the-holocaust"
     When I click on the author
     And I click on "Report User" from the content menu in the user info box
     And I confirm the reporting dialog because he is a holocaust denier:
       """
-      Do you really want to report the user "David Irving"?
+      Do you really want to report the user "I'm gonna mute Moderators and …"?
       """
-    Then I see a success message:
-      """
-      Thanks for reporting!
-      """
+    Then I see a toaster with "Thanks for reporting!"
 
   Scenario: Review reported content
     Given somebody reported the following posts:
       | submitterEmail           | resourceId | reasonCategory     | reasonDescription |
       | p1.submitter@example.org | p1         | discrimination_etc | Offensive content |
-    And I am logged in with a "moderator" role
+    And I am logged in as "moderator"
+    And I navigate to page "/"
     When I click on the avatar menu in the top right corner
     And I click on "Moderation"
     Then I see all the reported posts including the one from above
@@ -62,7 +58,8 @@ Feature: Report and Moderate
     Given somebody reported the following posts:
       | submitterEmail           | resourceId | reasonCategory | reasonDescription |
       | p2.submitter@example.org | p2         | other          | Offensive content |
-    And I am logged in with a "moderator" role
+    And I am logged in as "moderator"
+    And I navigate to page "/"
     And there is an annoying user who has muted me
     When I click on the avatar menu in the top right corner
     And I click on "Moderation"
@@ -70,6 +67,7 @@ Feature: Report and Moderate
     And I can visit the post page
 
   Scenario: Normal user can't see the moderation page
-    Given I am logged in with a "user" role
+    Given I am logged in as "user"
+    And I navigate to page "/"
     When I click on the avatar menu in the top right corner
     Then I can't see the moderation menu item
