+++ cypress/integration/Moderation.ReportContent/I_can_t_see_the_moderation_menu_item.js
@@ -1,0 +1,11 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then(`I can't see the moderation menu item`, () => {
+  cy.get('.avatar-menu-popover')
+    .find('a[href="/settings"]', 'Settings')
+    .should('exist') // OK, the dropdown is actually open
+  
+  cy.get('.avatar-menu-popover')
+    .find('a[href="/moderation"]', 'Moderation')
+    .should('not.exist')
+})
