+++ cypress/integration/Moderation.ReportContent/I_can_visit_the_post_page.js
@@ -1,0 +1,7 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then('I can visit the post page', () => {
+  cy.contains('Fake news').click()
+  cy.location('pathname').should('contain', '/post')
+    .get('.base-card .title').should('contain', 'Fake news')
+})
