+++ cypress/integration/Moderation.ReportContent/I_click_on_the_author.js
@@ -1,0 +1,7 @@
+import { When } from "cypress-cucumber-preprocessor/steps";
+
+When('I click on the author', () => {
+  cy.get('.user-teaser')
+    .click()
+    .url().should('include', '/profile/')
+})
