+++ cypress/integration/Moderation.ReportContent/I_click_on_the_avatar_menu_in_the_top_right_corner.js
@@ -1,0 +1,5 @@
+import { When } from "cypress-cucumber-preprocessor/steps";
+
+When("I click on the avatar menu in the top right corner", () => {
+  cy.get(".avatar-menu").click();
+});
