+++ cypress/integration/Moderation.ReportContent/I_see_all_the_reported_posts_including_from_the_user_who_muted_me.js
@@ -1,0 +1,7 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then('I see all the reported posts including from the user who muted me', () => {
+  cy.get('table tbody').within(() => {
+    cy.contains('tr', 'Fake news')
+  })
+})
