+++ cypress/integration/Moderation.ReportContent/I_see_all_the_reported_posts_including_the_one_from_above.js
@@ -1,0 +1,7 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then('I see all the reported posts including the one from above', () => {
+  cy.get('table tbody').within(() => {
+    cy.contains('tr', 'The Truth about the Holocaust')
+  })
+})
