+++ cypress/integration/Moderation.ReportContent/each_list_item_links_to_the_post_page.js
@@ -1,0 +1,6 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then('each list item links to the post page', () => {
+  cy.contains('The Truth about the Holocaust').click();
+  cy.location('pathname').should('contain', '/post')
+})
