+++ cypress/integration/Moderation.ReportContent/there_is_an_annoying_user_who_has_muted_me.js
@@ -1,0 +1,13 @@
+Given("there is an annoying user who has muted me", () => {
+  cy.neode()
+    .first("User", {
+      role: 'moderator'
+    })
+    .then(mutedUser => {
+      cy.neode()
+        .first("User", {
+          id: 'user'
+        })
+      .relateTo(mutedUser, "muted");
+    });
+});
